Moved DataSourceStream, CircularBuffer and OutputStream to SmingCore/Data.
