+++ Sming/SmingCore/Data/CircularBuffer.cpp
