+++ Sming/SmingCore/Data/Stream/DataSourceStream.h
@@ -5,8 +5,8 @@
  * All files of the Sming Core are provided under the LGPL v3 license.
  ****/
 
-#ifndef _SMING_CORE_DATASTREAM_H_
-#define _SMING_CORE_DATASTREAM_H_
+#ifndef _SMING_CORE_DATA_DATASTREAM_H_
+#define _SMING_CORE_DATA_DATASTREAM_H_
 
 #include <user_config.h>
 #include "../SmingCore/FileSystem.h"
@@ -368,4 +368,4 @@ private:
 };
 
 /** @} */
-#endif /* _SMING_CORE_DATASTREAM_H_ */
+#endif /* _SMING_CORE_DATA_DATASTREAM_H_ */
