+++ Sming/SmingCore/Network/Http/HttpConnection.h
@@ -17,7 +17,7 @@
 #include "HttpResponse.h"
 #include "HttpRequest.h"
 #include "../TcpClient.h"
-#include "../../DataSourceStream.h"
+#include "Data/Stream/DataSourceStream.h"
 #include "../../Services/DateTime/DateTime.h"
 
 typedef SimpleConcurrentQueue<HttpRequest*, HTTP_REQUEST_POOL_SIZE> RequestQueue;
